Move help database builder to new project
