+++ recipes/crud_site.gwr
@@ -9,7 +9,7 @@ web app setup:
     --project web.nav --home style-switcher
     --project web.site --home reader
     --project sql.crud --home posts?table=posts&dbfile=work/blog.sqlite
-web site build-help-db
+help-db build
 web:
  - static collect
  - server start-app --port 8888
