+++ recipes/etron/local.gwr
@@ -11,7 +11,7 @@ web app setup:
     - monitor.nmcli 
     - monitor.rpi --home pi-remote
     - ocpp.data --home charger-summary
-web site build-help-db
+help-db build
 
 # Toggle this version to apply the allowlist
 # ocpp csms setup-app --allow data/etron/rfids.cdv --location porsche_centre
