+++ recipes/gamebox.gwr
@@ -6,7 +6,7 @@ web app setup:
     - games --home games --links game-of-life,divination-wars,qpig-farm,massive-snake,fantastic-client
     - web.nav --home style-switcher
     - web.cookies --home cookie-jar
-web site build-help-db
+help-db build
 
 web:
  - static collect
