+++ recipes/skeleton.gwr
@@ -4,7 +4,7 @@ web app setup:
     - web.site --home reader
     - web.nav --style random
     - cdv --home data-editor
-web site build-help-db
+help-db build
 web:
  - static collect
 - server start-app
